+++ src/main/java/net/yxiao233/realmofdestiny/ModRegistry/ModItems.java
@@ -6,6 +6,8 @@ import net.minecraftforge.registries.DeferredRegister;
 import net.minecraftforge.registries.ForgeRegistries;
 import net.minecraftforge.registries.RegistryObject;
 import net.yxiao233.realmofdestiny.Items.ChangeStoneItem;
+import net.yxiao233.realmofdestiny.Items.CreativeEnergyMatrixItem;
+import net.yxiao233.realmofdestiny.Items.PedestalItem;
 import net.yxiao233.realmofdestiny.RealmOfDestiny;
 
 public class ModItems {
@@ -18,7 +20,7 @@ public class ModItems {
     public static final RegistryObject<Item> GEM_POLISHING_STATION_ITEM = ITEMS.register("gem_polishing_station",() ->
             new BlockItem(ModBlocks.GEM_POLISHING_STATION.get(),new Item.Properties()));
     public static final RegistryObject<Item> PEDESTAL_ITEM = ITEMS.register("pedestal", () ->
-            new BlockItem(ModBlocks.PEDESTAL.get(),new Item.Properties()));
+            new PedestalItem(ModBlocks.PEDESTAL.get(),new Item.Properties()));
 
     public static final RegistryObject<Item> BOLT_LOG_ITEM = ITEMS.register("bolt_log", () ->
             new BlockItem(ModBlocks.BOLT_LOG.get(),new Item.Properties()));
@@ -36,5 +38,5 @@ public class ModItems {
             new BlockItem(ModBlocks.BOLT_PLANKS.get(),new Item.Properties()));
 
     public static final RegistryObject<Item> CREATIVE_ENERGY_MATRIX_ITEM = ITEMS.register("creative_energy_matrix", () ->
-            new BlockItem(ModBlocks.CREATIVE_ENERGY_MATRIX.get(),new Item.Properties()));
+            new CreativeEnergyMatrixItem(ModBlocks.CREATIVE_ENERGY_MATRIX.get(),new Item.Properties()));
 }